+++ CHANGELOG.rst
@@ -11,6 +11,14 @@ Features
 * `#151 <https://github.com/readthedocs/sphinx-autoapi/issues/151>`: (Python) Added the ``autoapi_python_use_implicit_namespaces`` option to allow
   AutoAPI to search for implicit namespace packages.
 
+Bug Fixes
+^^^^^^^^^
+* `#186 <https://github.com/readthedocs/sphinx-autoapi/issues/186>`: (Python)
+  Fixed an exception when there are too many argument type annotations
+  in a type comment.
+  Instead, a warning is raised to indicate that the extra annotations
+  will be ignored.
+
 
 v1.2.1 (2019-10-9)
 ------------------